run as background task, change url, add admin panel
