+++ beagle_etl/views.py
@@ -1,13 +1,38 @@
+from django.conf import settings
 from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.generics import GenericAPIView
-from beagle_etl.models import ETLConfiguration
+from rest_framework.views import APIView
+from rest_framework.permissions import IsAuthenticated
+from beagle_etl.models import ETLConfiguration, SMILEMessage
+from beagle_etl.jobs.metadb_jobs import new_request
 from drf_yasg.utils import swagger_auto_schema
 from .serializers import (
     AssaySerializer,
     AssayElementSerializer,
     AssayUpdateSerializer,
 )
+
+
+class ForceImportView(APIView):
+    permission_classes = (IsAuthenticated,)
+
+    def post(self, _request, request_id):
+        message = (
+            SMILEMessage.objects.filter(request_id=request_id, topic=settings.METADB_NATS_NEW_REQUEST)
+            .order_by("-created_date")
+            .first()
+        )
+        if not message:
+            return Response(
+                {"detail": f"No new-request SMILEMessage found for request_id {request_id}."},
+                status=status.HTTP_404_NOT_FOUND,
+            )
+        new_request.delay(str(message.id), force_import=True)
+        return Response(
+            {"detail": f"Force import triggered for request {request_id} (message {message.id})."},
+            status=status.HTTP_202_ACCEPTED,
+        )
 
 
 class AssayViewSet(GenericAPIView):
